Create Memory_game.java
The Memory Game is a game where the player has to flip over pairs of cards with the same symbol.

We create two arrays to represent the game board: board and flipped. The board is an array of strings that represents the state of the game board at any given time.
When a player flips a card, we replace the corresponding element of the board array with the value of the card.
Flipped is a boolean array that keeps track of which cards have been flipped over.
All elements of the flipped array are false. When a player flips a card, we set the corresponding element of the flipped array to true.